requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0 
python-dotenv>=0.19.0 
scipy>=1.7.0
humanize>=4.0.0
datasets>=2.0.0
huggingface_hub>=0.16.0
PyYAML>=6.0.0